--- a/R/dfm.R
+++ b/R/dfm.R
@@ -313,11 +313,11 @@ dfm.dfm <- function(x,
     
     if (!is.null(groups)) {
         if (is.character(groups) & all(groups %in% names(docvars(x)))) {
-            groups <- as.factor(interaction(docvars(x)[, groups], drop = TRUE))
+            groups <- as.charcter(interaction(docvars(x)[, groups], drop = TRUE))
         } else {
             if (length(groups) != ndoc(x))
                 stop("groups must name docvars or provide data matching the documents in x")
-            groups <- as.factor(groups)
+            groups <- as.charcter(groups)
         }
         if (verbose)
             catm("   ... grouping texts\n")