--- a/man/bootstrap_dfm.Rd
+++ b/man/bootstrap_dfm.Rd
@@ -31,7 +31,7 @@ is done at the sentence level, within document.
 \examples{
 txt <- c(textone = "This is a sentence.  Another sentence.  Yet another.", 
          texttwo = "Premiere phrase.  Deuxieme phrase.")
-bootstrap_dfm(txt, n = 3)         
+bootstrap_dfm(txt, n = 3)    
 }
 \author{
 Kenneth Benoit