I added `tokens_group()` to group documents before making dfm for #813. It is used when groups is given to `dfm()`)

```r
> toks <- readRDS('data_tokens_sydney_v2.RDS')
> 
> print(object.size(toks), units = 'MB')
1332.9 Mb
> 
> microbenchmark::microbenchmark(
+     new = dfm(toks, groups = docvars(toks, 'year'), tolower = FALSE),
+     old = dfm_group(dfm(toks, tolower = FALSE), groups = docvars(toks, 'year')),
+     times = 10
+ )
Unit: seconds
 expr       min        lq     mean    median       uq      max neval
  new  9.233766  9.471829 10.68980  9.831215 10.88419 16.89836    10
  old 11.011277 11.133569 11.40997 11.318823 11.47242 12.59041    10
> 
> null <- gc(reset = TRUE)
> dfm(toks, groups = docvars(toks, 'year'), tolower = FALSE)
Document-feature matrix of: 31 documents, 362,824 features (80% sparse).
> gc()
            used  (Mb) gc trigger   (Mb)  max used   (Mb)
Ncells   9770401 521.8   14442815  771.4  14442815  771.4
Vcells 117340537 895.3  559296049 4267.1 694770055 5300.7
> 
> null <- gc(reset = TRUE)
> dfm_group(dfm(toks, tolower = FALSE), groups = docvars(toks, 'year'))
Document-feature matrix of: 31 documents, 362,824 features (80% sparse).
> gc()
            used  (Mb) gc trigger   (Mb)  max used   (Mb)
Ncells   9771046 521.9   14442815  771.4   9989286  533.5
Vcells 117343054 895.3  632020472 4822.0 779315158 5945.8
```