@koheiw 2017-06-23
My intention to add `tokens_group()` is pre-group document in the same way as dfm does with characters with `texts()`. `tokens_group()` can be exported if we add padding at the end of the each documents to avoid false ngrams, but kept it internal only for now.

However, admittedly, use of `tokens_group()` is not the most efficient way to group documents: it is better to re-indexing groups just before making a matrix with `sparseMatrix()`. Yet, it is difficult to optimize dfm construction without major changes, because it calls dfm.dfm(). We have to simplify the process.