@kbenoit 2017-06-27
@koheiw to strip tokens_group code and tests out, start a new branch and PR. Make the functionality and documentation as close to dfm_group as possible, but do not export, and mark the documentation as @keyword internal.